Machine
  A laptop (also laptop computer), often called a notebook, is a small, portable personal computer (PC) with a "clamshell" form factor, typically having a thin LCD or LED computer screen mounted on the inside of the upper lid of the clamshell and an alphanumeric keyboard on the inside of the lower lid. Today, laptops are commonly used in a variety of settings, such as at work, in education, for playing games, Internet surfing, for personal multimedia, and general home computer use.
Histroy: As the personal computer (PC) became feasible in 1971. A "personal, portable information manipulator" was imagined by Alan Kay at Xerox PARC in 1968,  and described in his 1972 paper as the "Dynabook" The IBM Special Computer APL Machine Portable (SCAMP) was demonstrated in 1973. This prototype was based on the IBM PALM processor.  The IBM 5100 in September 1975. 
My Laptop Specifications: